MODIFY:   change the binding for xs:nonNegativeInteger (used in the schema files)   from Integer to BigInteger (fixes bug #1475274) and provide a Digit   enum to be used in NumStroke (to avoid using BigInteger there as well).

+++ parser/src/net/sourceforge/czt/parser/util/LocInfoImpl.java
@@ -37,10 +37,10 @@ public class LocInfoImpl
   {
     if (locAnn != null) {
       source_ = locAnn.getLoc();
-      if (locAnn.getLine() != null) line_ = locAnn.getLine();
-      if (locAnn.getCol() != null) column_ = locAnn.getCol();
-      if (locAnn.getStart() != null) start_ = locAnn.getStart();
-      if (locAnn.getLength() != null) length_ = locAnn.getLength();
+      if (locAnn.getLine() != null) line_ = locAnn.getLine().intValue();
+      if (locAnn.getCol() != null) column_ = locAnn.getCol().intValue();
+      if (locAnn.getStart() != null) start_ = locAnn.getStart().intValue();
+      if (locAnn.getLength() != null) length_ = locAnn.getLength().intValue();
     }
   }
